fix: cast default bus flags to u32 in vst3 example

+++ examples/minimal-vst3/src/processor.rs
@@ -111,7 +111,7 @@ impl IComponentTrait for GainProcessor {
         bus.channelCount = 2;
         copy_wstring(label, &mut bus.name);
         bus.busType = BusTypes_::kMain as BusType;
-        bus.flags = BusInfo_::BusFlags_::kDefaultActive;
+        bus.flags = BusInfo_::BusFlags_::kDefaultActive as u32;
 
         kResultOk
     }
